moved JobValidatorService
